+++ test/test_parser_options.gd
@@ -16,19 +16,22 @@ npc: what do you want to talk about?
 * Everything else... #some_tag
 	player: What about everything else?
 	npc: I don't have time for this...
+* one more thing $abc&whatever
+	npc: one
 """ )
 	var expected = {
 		"type": 'document',
 		"content": [{
 			"type": 'content',
 			"content": [
-				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null },
+				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 				{
 					"type": 'options',
 					"name": null,
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
@@ -36,12 +39,13 @@ npc: what do you want to talk about?
 							"speaker": 'speaker',
 							"id": null,
 							"tags": null,
+							"id_suffixes": null,
 							"mode": 'once',
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, },
-									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -54,12 +58,28 @@ npc: what do you want to talk about?
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null },
-									{ "type": 'line', "value": 'I don\'t have time for this...', "speaker": 'npc', "id": null, "tags": null },
+									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'I don\'t have time for this...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 							"tags": [ 'some_tag', ],
+							"id_suffixes": null,
 						},
+						{
+							"type": "option",
+							"name": "one more thing",
+							"mode": "once",
+							"speaker": null,
+							"tags": null,
+							"content": {
+								"type": "content",
+								"content": [
+										{ "type": "line", "value": "one", "speaker": "npc","tags": null, "id": null, "id_suffixes": null, },
+									],
+								},
+							"id": "abc",
+							"id_suffixes": [ "whatever" ],
+							},
 					],
 				},
 			],
@@ -83,13 +103,14 @@ npc: what do you want to talk about?
 		"content": [{
 			"type": 'content',
 			"content": [
-				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null },
+				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 				{
 					"type": 'options',
 					"name": null,
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
@@ -98,10 +119,11 @@ npc: what do you want to talk about?
 							"speaker": null,
 							"id": null,
 							"tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -114,10 +136,11 @@ npc: what do you want to talk about?
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null },
+									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 							"tags": [ 'some_tag', ],
+							"id_suffixes": null,
 						},
 					],
 				},
@@ -142,13 +165,14 @@ npc: what do you want to talk about?
 		"content": [{
 			"type": 'content',
 			"content": [
-				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null },
+				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 				{
 					"type": 'options',
 					"name": null,
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
@@ -157,10 +181,11 @@ npc: what do you want to talk about?
 							"speaker": null,
 							"id": null,
 							"tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -173,10 +198,11 @@ npc: what do you want to talk about?
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null },
+									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 							"tags": [ 'some_tag', ],
+							"id_suffixes": null,
 						},
 					],
 				},
@@ -204,25 +230,27 @@ npc: what do you want to talk about?
 		"content": [{
 			"type": 'content',
 			"content": [
-				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null },
+				{ "type": 'line', "value": 'what do you want to talk about?', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 				{
 					"type": 'options',
 					"name": null,
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
 							"name": 'Life',
 							"mode": 'once',
 							"id": null, "tags": null, "speaker": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'Life', "speaker": null, "id": null, "tags": null, },
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, },
-									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'Life', "speaker": null, "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -230,12 +258,13 @@ npc: what do you want to talk about?
 							"type": 'option',
 							"name": 'Everything else...',
 							"mode": 'once', "id": null, "speaker": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'Everything else...', "speaker": null, "id": null, "tags": [ 'some_tag', ], },
-									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null, },
-									{ "type": 'line', "value": 'I don\'t have time for this...', "speaker": 'npc', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'Everything else...', "speaker": null, "id": null, "tags": [ 'some_tag', ], "id_suffixes": null, },
+									{ "type": 'line', "value": 'What about everything else?', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'I don\'t have time for this...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 							"tags": [ 'some_tag', ],
@@ -255,7 +284,8 @@ func test_use_first_line_as_label():
 	life
 	player: I want to talk about life!
 	npc: Well! That's too complicated...
-
+*
+	the universe #tag $id&suffix
 """ )
 	var expected = {
 		"type": 'document',
@@ -268,17 +298,31 @@ func test_use_first_line_as_label():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
 							"name": 'life',
 							"mode": 'once', "id": null, "tags": null, "speaker": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'life', "id": null, "speaker": null, "tags": null },
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null },
-									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null },
+									{ "type": 'line', "value": 'life', "id": null, "speaker": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
+								],
+							},
+						},
+						{
+							"type": 'option',
+							"name": 'the universe',
+							"mode": 'once', "id": "id", "tags": ["tag"], "speaker": null,
+							"id_suffixes": ["suffix"],
+							"content": {
+								"type": 'content',
+								"content": [
+									{ "type": 'line', "value": "the universe", "id": "id", "speaker": null, "tags": ["tag"], "id_suffixes": ["suffix"], },
 								],
 							},
 						},
@@ -294,7 +338,7 @@ func test_use_first_line_as_label():
 
 func test_use_previous_line_as_label():
 	var result = parse("""
-spk: this line will be the label $some_id #some_tag
+spk: this line will be the label $some_id&some_suffix #some_tag
 	* life
 		player: I want to talk about life!
 		npc: Well! That's too complicated...
@@ -313,17 +357,19 @@ spk: second try
 					"speaker": 'spk',
 					"id": 'some_id',
 					"tags": ['some_tag'],
+					"id_suffixes": ["some_suffix"],
 					"name": 'this line will be the label',
 					"content": [
 						{
 							"type": 'option',
 							"name": 'life',
 							"mode": 'once', "id": null, "speaker": null, "tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, },
-									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
+									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -335,15 +381,17 @@ spk: second try
 					"name": 'second try',
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
 							"name": 'life',
 							"mode": 'once', "id": null, "speaker": null, "tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'Well! That\'s too complicated...', "speaker": 'npc', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -377,16 +425,18 @@ func test_use_previous_line_in_quotes_as_label():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"name": 'spk: this line will be the label $some_id #some_tag',
 					"content": [
 						{
 							"type": 'option',
 							"name": 'life',
 							"mode": 'once', "id": null, "tags": null, "speaker": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'I want to talk about life!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -396,6 +446,7 @@ func test_use_previous_line_in_quotes_as_label():
 					"type": 'options',
 					"name": 'spk: this line will be the label $some_id #some_tag',
 					"tags": null,
+					"id_suffixes": null,
 					"speaker": null,
 					"id": null,
 					"content": [
@@ -403,10 +454,11 @@ func test_use_previous_line_in_quotes_as_label():
 							"type": 'option',
 							"name": 'universe',
 							"mode": 'once', "id": null, "speaker": null, "tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'I want to talk about the universe!', "speaker": 'player', "id": null, "tags": null, },
+									{ "type": 'line', "value": 'I want to talk about the universe!', "speaker": 'player', "id": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -438,16 +490,18 @@ func test_ensures_options_ending_worked():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": 'option',
 							"name": 'yes',
 							"mode": 'once',
 							"id": null, "speaker": null, "tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null },
+									{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -456,10 +510,11 @@ func test_ensures_options_ending_worked():
 							"name": 'no',
 							"mode": 'once',
 							"id": null, "speaker": null, "tags": null,
+							"id_suffixes": null,
 							"content": {
 								"type": 'content',
 								"content": [
-									{ "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null },
+									{ "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null, "id_suffixes": null, },
 								],
 							},
 						},
@@ -468,7 +523,7 @@ func test_ensures_options_ending_worked():
 				{
 					"type": "conditional_content",
 					"conditions": { "type": "variable", "name": "some_check" },
-					"content": { "type": "line", "value": "maybe", "id": null, "speaker": null, "tags": null }
+					"content": { "type": "line", "value": "maybe", "id": null, "speaker": null, "tags": null, "id_suffixes": null, }
 				},
 			],
 		},
@@ -495,6 +550,7 @@ func test_ensures_option_item_ending_worked():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": "action_content",
@@ -512,15 +568,15 @@ func test_ensures_option_item_ending_worked():
 							"content": {
 								"type": 'option',
 								"name": 'yes',
-								"mode": 'once', "id": null, "speaker": null, "tags": null,
-								"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null }]},
+								"mode": 'once', "id": null, "speaker": null, "tags": null, "id_suffixes": null,
+								"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }]},
 							},
 						},
 						{
 							"type": 'option',
 							"name": 'no',
-							"mode": 'once', "id": null, "speaker": null, "tags": null,
-							"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no' , "id": null, "speaker": null, "tags": null}, ]},
+							"mode": 'once', "id": null, "speaker": null, "tags": null, "id_suffixes": null,
+							"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no' , "id": null, "speaker": null, "tags": null, "id_suffixes": null, }, ]},
 						},
 					],
 				},
@@ -548,6 +604,7 @@ func test_options_with_blocks_both_sides():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 						 "type": "conditional_content",
@@ -569,7 +626,8 @@ func test_options_with_blocks_both_sides():
 									"type": 'option',
 									"name": 'yes',
 									"mode": 'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null }]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }]},
 								},
 							},
 					 },
@@ -594,7 +652,8 @@ func test_options_with_blocks_both_sides():
 									"type": 'option',
 									"name": 'no',
 									"mode": 'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null }, ]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }, ]},
 								},
 							},
 						},
@@ -626,6 +685,7 @@ func test_options_with_multiple_blocks_on_same_side():
 					"speaker": null,
 					"id": null,
 					"tags": null,
+					"id_suffixes": null,
 					"content": [
 						{
 							"type": "conditional_content",
@@ -647,7 +707,8 @@ func test_options_with_multiple_blocks_on_same_side():
 									"type": 'option',
 									"name": 'yes',
 									"mode":  'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null }]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }]},
 								},
 							},
 						},
@@ -672,7 +733,8 @@ func test_options_with_multiple_blocks_on_same_side():
 									"type": 'option',
 									"name": 'no',
 									"mode":  'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null }, ]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }, ]},
 								},
 							},
 						},
@@ -697,7 +759,8 @@ func test_options_with_multiple_blocks_on_same_side():
 									"type": 'option',
 									"name": 'yes',
 									"mode":  'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null }]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [{ "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }]},
 								},
 							},
 						},
@@ -722,7 +785,8 @@ func test_options_with_multiple_blocks_on_same_side():
 									"type": 'option',
 									"name": 'no',
 									"mode":  'once', "id": null, "speaker": null, "tags": null,
-									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null }, ]},
+									"id_suffixes": null,
+									"content": { "type": 'content', "content": [ { "type": 'line', "value": 'no', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }, ]},
 								},
 							},
 						},
@@ -760,7 +824,8 @@ func test_options_with_multiple_blocks_on_same_side():
 										"type": 'option',
 										"name": 'yes',
 										"mode":  'once', "id": null, "speaker": null, "tags": null,
-										"content": { "type": 'content', "content": [ { "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null }, ]},
+										"id_suffixes": null,
+										"content": { "type": 'content', "content": [ { "type": 'line', "value": 'yes', "id": null, "speaker": null, "tags": null, "id_suffixes": null, }, ]},
 									},
 								},
 							},