+++ CHANGELOG.md
@@ -2,11 +2,18 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unreleased
+
+## Added
+
+- Implemented support for id suffixes. `dialogue line $id&variable_1&variable_2`.
+- Id Suffix Lookup Separator project setttings. `Project > Project Settings > General > Dialogue > Id Suffix Lookup Separator`.
+
 ## 3.1.0
 
 ## Added
 
-- Dialogue source folder project setttings. `Project > Project Settings > Dialogue > Source Folder`.
+- Dialogue source folder project setttings. `Project > Project Settings > General > Dialogue > Source Folder`.
   When only the file name is provided, the interpreter will look for files in the folder set on `dialogue/source_folder`.
 
 ## Fixed